requirements.txt
# Core dependencies
python-dotenv==1.1.1
openai==1.101.0
requests==2.32.3

# Phoenix tracing
arize-phoenix-otel==0.13.0
openinference-instrumentation-openai==0.1.31
openinference-instrumentation-langchain==0.1.50

# LangChain and LangGraph
langchain==0.3.27
langchain-openai==0.3.31
langgraph==0.6.6

# Testing
pytest==8.4.1
pytest-asyncio==1.1.0